+++ release-notes.md
@@ -1,3 +1,12 @@
+Pending changes (not released yet)
+==================================
+
+Changes:
+  - New configuration "Tokens.LandingOptions": when opening a share link, this option enables
+    the display of a landing page with, e.g., a download button, a button to open the study in
+    the viewer and, optionally, a custom button.
+     
+
 1.8.3 (2025-04-07)
 ==================